Updates setup script to not build dev js on prod

+++ bin/setup.sh
@@ -1,10 +1,13 @@
 mkdir public
 mkdir public/js
 npm run build
-{
-  # npm run build watches for changes, so this kills it after the assets has def. finished building
-  sleep 5s
-  echo 'Done!'
-  kill $$
-} &
-npm run build_dev
+if [ -n "PRODUCTION" ]
+then
+  {
+    # npm run build watches for changes, so this kills it after the assets has def. finished building
+    sleep 5s
+    echo 'Done!'
+    kill $$
+  } &
+  npm run build_dev;
+fi
